3 warnings left i think

+++ src/scenario/event/parameter_data.c
@@ -352,7 +352,7 @@ static scenario_action_data_t scenario_action_data[ACTION_TYPE_MAX] = {
     [ACTION_TYPE_CHANGE_PRODUCTION_RATE] = {.type = ACTION_TYPE_CHANGE_PRODUCTION_RATE,
                                         .xml_attr = {.name = "change_production_rate",  .type = PARAMETER_TYPE_TEXT,    .key = TR_ACTION_TYPE_PRODUCTION_RATE},
                                         .xml_parm1 = {.name = "resource",       .type = PARAMETER_TYPE_RESOURCE,   .key = TR_PARAMETER_TYPE_RESOURCE },
-                                        .xml_parm2 = {.name = "rate",         .type = PARAMETER_TYPE_FORMULA,            .min_limit = 0,           
+                                        .xml_parm2 = {.name = "rate",         .type = PARAMETER_TYPE_FORMULA,            .min_limit = 0,
                                             .max_limit = UNLIMITED,     .key = TR_PARAMETER_TYPE_NUMBER },
                                         .xml_parm3 = {.name = "set_to_value",      .type = PARAMETER_TYPE_BOOLEAN,      .min_limit = 0,
                                             .max_limit = 1,         .key = TR_PARAMETER_SET_TO_VALUE }, }
@@ -1210,11 +1210,11 @@ const uint8_t *scenario_events_parameter_data_get_display_string(special_attribu
             }
             break;
         case PARAMETER_TYPE_GRID_SLICE:
-            {
+        {
             static uint8_t buffer[16];
             string_from_int(buffer, entry->value, 0);
             return (const uint8_t *) buffer;
-            }
+        }
         default:
             return translation_for(entry->key);
     }
@@ -1249,7 +1249,7 @@ static uint8_t *string_from_year(uint8_t *dst, int year, int *maxlength)
 
 static uint8_t *translation_for_request_value(int value, uint8_t *result_text, int *maxlength)
 {
-    if (value < 0 || value >= scenario_request_count_total()) {
+    if (value < 0 || value >= (int) scenario_request_count_total()) {
         return string_copy(translation_for(TR_PARAMETER_VALUE_NONE), result_text, *maxlength);
     }
     const scenario_request *request = scenario_request_get(value);
